mindmate_bot/requirements.txt
python-telegram-bot==20.7
openai==1.3.0
transformers==4.35.0
torch==2.1.0 --index-url https://download.pytorch.org/whl/cpu
fastapi==0.104.0
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.3
scikit-learn==1.3.2
numpy==1.24.3
pydantic==2.5.0
aiofiles==23.2.1